refactor: dto folder structure (#19)
