+++ src/main/java/app/pigrest/auth/controller/LoginController.java
@@ -1,12 +1,5 @@
 package app.pigrest.auth.controller;
 
-import app.pigrest.auth.dto.LoginRequest;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
